fix msvc issues

+++ database_structures.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 
 namespace database
@@ -7,27 +8,27 @@ namespace database
 
 struct group
 {
-    ssize_t id;
+    intmax_t id;
     std::string name;
 };
 
 struct permission
 {
-    ssize_t id;
+    intmax_t id;
     std::string name;
 };
 
 struct group_permission
 {
-    ssize_t id;
-    ssize_t group_id;
-    ssize_t permission_id;
+    intmax_t id;
+    intmax_t group_id;
+    intmax_t permission_id;
 };
 
 struct user
 {
-    ssize_t id;
-    ssize_t group_id;
+    intmax_t id;
+    intmax_t group_id;
     std::string name;
     std::string email;
     std::string pass;
@@ -35,31 +36,31 @@ struct user
 
 struct project
 {
-    ssize_t id;
+    intmax_t id;
     std::string name;
     bool deleted;
 };
 
 struct token
 {
-    ssize_t id;
-    ssize_t user_id;
+    intmax_t id;
+    intmax_t user_id;
     std::string value;
     uint64_t expiration_time;
 };
 
 struct api_access_key
 {
-    ssize_t id;
-    ssize_t user_id;
+    intmax_t id;
+    intmax_t user_id;
     std::string value;
 };
 
 struct project_user_right
 {
-    ssize_t id;
-    ssize_t project_id;
-    ssize_t user_id;
+    intmax_t id;
+    intmax_t project_id;
+    intmax_t user_id;
     bool read;
     bool write;
     bool execute;
